Imported upstream version '3.26.3' of 'upstream'

+++ joint_trajectory_controller/include/joint_trajectory_controller/tolerances.hpp
@@ -293,7 +293,7 @@ SegmentTolerances get_segment_tolerances(
  * REALTIME if true \return True if \p state_error fulfills \p state_tolerance.
  */
 inline bool check_state_tolerance_per_joint(
-  const trajectory_msgs::msg::JointTrajectoryPoint & state_error, int joint_idx,
+  const trajectory_msgs::msg::JointTrajectoryPoint & state_error, size_t joint_idx,
   const StateTolerances & state_tolerance, bool show_errors = false)
 {
   using std::abs;
@@ -316,7 +316,7 @@ inline bool check_state_tolerance_per_joint(
   if (show_errors)
   {
     const auto logger = rclcpp::get_logger("tolerances");
-    RCLCPP_ERROR(logger, "State tolerances failed for joint %d:", joint_idx);
+    RCLCPP_ERROR(logger, "State tolerances failed for joint %lu:", joint_idx);
 
     if (state_tolerance.position > 0.0 && abs(error_position) > state_tolerance.position)
     {
